Fix minimap chars-per-dot ignored in pixel mode, and accept fractional values
EnsurePlane() hardcoded charsPerDot=1 for real-pixel-mode rendering
regardless of ned/set-minimap-chars-per-dot's configured value -- the
setting only ever affected the glyph-mode fallback, so anyone actually
running in pixel mode (the common case on a capable terminal) saw zero
visual effect from changing it. Both render paths now uniformly honor
editor::MinimapCharsPerDot().

Also switches the setting from int to double throughout (storage, Janet
binding, the two arithmetic spots in ForEachDensityDot) so a value between
whole-number steps (e.g. 8.5) is reachable, not just coarser integer jumps.

+++ Source/Editor/MinimapSettings.cpp
@@ -22,8 +22,8 @@ namespace {
         return width;
     }
 
-    int& CharsPerDotStorage() {
-        static int charsPerDot = 8;
+    double& CharsPerDotStorage() {
+        static double charsPerDot = 8.0;
         return charsPerDot;
     }
 
@@ -49,12 +49,12 @@ int MinimapWidth() {
     return WidthStorage();
 }
 
-void SetMinimapCharsPerDot(int columns) {
+void SetMinimapCharsPerDot(double columns) {
     const std::lock_guard<std::mutex> lock(SettingsMutex());
-    CharsPerDotStorage() = std::max(1, columns);
+    CharsPerDotStorage() = std::max(1.0, columns);
 }
 
-int MinimapCharsPerDot() {
+double MinimapCharsPerDot() {
     const std::lock_guard<std::mutex> lock(SettingsMutex());
     return CharsPerDotStorage();
 }
